+++ nodeconductor/cost_tracking/__init__.py
@@ -25,6 +25,7 @@ import logging
 from decimal import Decimal
 
 from django.contrib.contenttypes.models import ContentType
+from nodeconductor.structure import ServiceBackendNotImplemented
 
 
 default_app_config = 'nodeconductor.cost_tracking.apps.CostTrackingConfig'
@@ -46,7 +47,10 @@ class CostTrackingRegister(object):
 
     @classmethod
     def get_resource_backend(cls, resource):
-        return cls._register.get(resource._meta.app_label)
+        try:
+            return cls._register[resource._meta.app_label]
+        except KeyError:
+            raise ServiceBackendNotImplemented
 
 
 class CostTrackingBackend(object):
@@ -87,7 +91,7 @@ class CostTrackingBackend(object):
         monthly_cost = 0
         for item_type, item_key, item_count in cls.get_used_items(resource):
             try:
-                monthly_cost += resource_prices[(item_type, item_key)] * item_count
+                monthly_cost += resource_prices[(item_type, item_key)] * Decimal(format(item_count, ".15g"))
             except KeyError:
                 logger.error('Can not find price item with key "%s" and type "%s" for resource "%s"',
                              item_key, item_type, resource_content_type.name)