+++ nodeconductor/template/models.py
@@ -3,6 +3,8 @@ import json
 from celery import chain
 from django import template as django_template
 from django.contrib.contenttypes.models import ContentType
+from django.core import validators
+from django.core.exceptions import ValidationError
 from django.db import models
 from django.utils.encoding import python_2_unicode_compatible
 from jsonfield import JSONField
@@ -13,6 +15,7 @@ from rest_framework.authtoken.models import Token
 from taggit.managers import TaggableManager
 
 from nodeconductor.core import models as core_models
+from nodeconductor.structure import models as structure_models
 
 
 @python_2_unicode_compatible
@@ -110,6 +113,7 @@ class TemplateActionException(Exception):
             return {'message': str(serialized_exception)}
 
 
+@python_2_unicode_compatible
 class Template(core_models.UuidMixin, models.Model):
     """ Template for application action.
 
@@ -118,11 +122,15 @@ class Template(core_models.UuidMixin, models.Model):
     """
     group = models.ForeignKey(TemplateGroup, related_name='templates')
     options = JSONField(default={}, help_text='Default options for resource provision request.')
+    tags = TaggableManager()
+    service_settings = models.ForeignKey(structure_models.ServiceSettings, related_name='templates', null=True)
     resource_content_type = models.ForeignKey(
         ContentType, help_text='Content type of resource which provision process is described in template.')
     order_number = models.PositiveSmallIntegerField(
-        default=1, help_text='Templates in group are sorted by order number. '
-                             'Template with smaller order number will be executed first.')
+        default=1,
+        help_text='Templates in group are sorted by order number. '
+                  'Template with smaller order number will be executed first.',
+        validators=[validators.MinValueValidator(1)])
     use_previous_resource_project = models.BooleanField(
         default=False, help_text='If True and project is not defined in template - current resource will use the same '
                                  'project as previous created.')
@@ -205,6 +213,14 @@ class Template(core_models.UuidMixin, models.Model):
                       response.request.url, response.status_code, response.content)
             raise TemplateActionException(message, details, response.status_code)
         return response
+
+    def clean(self):
+        if self.group.templates.count() < self.order_number:
+            raise ValidationError(
+                {'order_number': 'Template order number cannot be greater than count of templates in group'})
+
+    def __str__(self):
+        return "%s -> %s" % (self.group.name, self.resource_content_type)
 
 
 class TemplateGroupResult(core_models.UuidMixin, TimeStampedModel):