+++ CHANGELOG.rst
@@ -3,6 +3,7 @@ CHANGELOG
 
 Release NEXT
 ------------
+- Add project filter to template groups
 - Added recovery transition from ERRED to SYNCING state
 
 Release 0.82.0